+++ src/db/index.ts
@@ -1,12 +1,12 @@
 import { DependencyRepositoryProvider } from './impl/dependencyRepositoryProvider';
-import { IDependencyRepositoryProvider } from './interfaces/dependencyRepositoryProvider';
+import { IDependencyRepositoryProvider } from './interfaces/IDependencyRepositoryProvider';
 import { ConnectionProvider } from './impl/connectionProvider';
 import { DependencyVersionRepositoryProvider } from './impl/dependencyVersionRepositoryProvider';
-import { IDependencyVersionRepositoryProvider } from './interfaces/dependencyVersionRepositoryProvider';
-import { IConnectionProvider } from './interfaces/connectionProvider';
+import { IDependencyVersionRepositoryProvider } from './interfaces/IDependencyVersionRepositoryProvider';
+import { IConnectionProvider } from './interfaces/IConnectionProvider';
 import { interfaces } from 'inversify';
 import Bind = interfaces.Bind;
-import { IEntity } from './interfaces/entity';
+import { IEntity } from './interfaces/IEntity';
 import { Dependency } from './entities/dependency';
 import { DependencyVersion } from './entities/dependencyVersion';
 
@@ -22,9 +22,9 @@ export const dbModulesBinder = (bind: Bind): void => {
   bind<IEntity>(IEntity).toConstantValue(DependencyVersion);
 };
 
-export * from './interfaces/connectionProvider';
-export * from './interfaces/dependencyRepositoryProvider';
-export * from './interfaces/dependencyVersionRepositoryProvider';
-export * from './interfaces/connectionSettings';
+export * from './interfaces/IConnectionProvider';
+export * from './interfaces/IDependencyRepositoryProvider';
+export * from './interfaces/IDependencyVersionRepositoryProvider';
+export * from './interfaces/IConnectionSettings';
 export * from './entities/dependency';
 export * from './entities/dependencyVersion';
