+++ src/db/interfaces/IConnectionSettings.ts
