+++ src/db/interfaces/IEntity.ts
