+++ src/flow/index.ts
@@ -1,4 +1,4 @@
-import { IFlow } from './interfaces/flow';
+import { IFlow } from './interfaces/IFlow';
 import { Flow } from './impl/flow';
 import { interfaces } from 'inversify';
 import Bind = interfaces.Bind;
@@ -7,4 +7,4 @@ export const flowModulesBinder = (bind: Bind): void => {
   bind<IFlow>(IFlow).to(Flow).inSingletonScope();
 };
 
-export * from './interfaces/flow';
+export * from './interfaces/IFlow';
