+++ src/resolvers/cacheResolver/index.ts
@@ -1,4 +1,4 @@
-import { ICacheResolver } from './interfaces/cacheResolver';
+import { ICacheResolver } from './interfaces/ICacheResolver';
 import { CacheResolver } from './impl/cacheResolver';
 import { interfaces } from 'inversify';
 import Bind = interfaces.Bind;
@@ -7,4 +7,4 @@ export const cacheResolveModulesBinder = (bind: Bind): void => {
   bind<ICacheResolver>(ICacheResolver).to(CacheResolver).inSingletonScope();
 };
 
-export * from './interfaces/cacheResolver';
+export * from './interfaces/ICacheResolver';
