+++ src/resolvers/cacheResolver/interfaces/ICacheResolver.ts
