+++ src/resolvers/ciResolver/interfaces/ICIResolver.ts
