+++ src/resolvers/ciResolver/interfaces/ISpecificCIResolver.ts
