+++ src/resolvers/enginesResolver/index.ts
@@ -1,10 +1,10 @@
 import { interfaces } from 'inversify';
 import Bind = interfaces.Bind;
-import { IEnginesResolver } from './interfaces/enginesResolver';
+import { IEnginesResolver } from './interfaces/IEnginesResolver';
 import { EnginesResolver } from './impl/enginesResolver';
 
 export const enginesResolveModulesBinder = (bind: Bind): void => {
   bind<IEnginesResolver>(IEnginesResolver).to(EnginesResolver).inSingletonScope();
 };
 
-export * from './interfaces/enginesResolver';
+export * from './interfaces/IEnginesResolver';
