+++ src/resolvers/testResolver/index.ts
@@ -1,5 +1,5 @@
 import { TestResolver } from './impl/testResolver';
-import { ITestResolver } from './interfaces/testResolver';
+import { ITestResolver } from './interfaces/ITestResolver';
 import { interfaces } from 'inversify';
 import Bind = interfaces.Bind;
 
@@ -7,4 +7,4 @@ export const testResolverModulesBinder = (bind: Bind): void => {
   bind<ITestResolver>(ITestResolver).to(TestResolver).inSingletonScope();
 };
 
-export * from './interfaces/testResolver';
+export * from './interfaces/ITestResolver';
