+++ src/utils/git/impl/git.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from 'inversify';
 import { SimpleGitFn, TYPES } from '../../../container/nodeModulesContainer';
 import { ILoggerFactory } from '../../logger';
-import { ILogger } from '../../logger/interfaces/logger';
+import { ILogger } from '../../logger/interfaces/ILogger';
 import { SimpleGit } from 'simple-git/promise';
 
 export interface ICheckoutCommitOptions {
