+++ src/utils/git/index.ts
@@ -1,7 +1,7 @@
 import { interfaces } from 'inversify';
 import { Git } from './impl/git';
 import { GitCheckout } from './impl/gitCheckout';
-import { IGitCheckout } from './interfaces/gitCheckout';
+import { IGitCheckout } from './interfaces/IGitCheckout';
 import Bind = interfaces.Bind;
 
 export const gitModuleBinder = (bind: Bind): void => {
@@ -9,4 +9,4 @@ export const gitModuleBinder = (bind: Bind): void => {
   bind<IGitCheckout>(IGitCheckout).to(GitCheckout).inSingletonScope();
 };
 
-export * from './interfaces/gitCheckout';
+export * from './interfaces/IGitCheckout';
