+++ src/utils/logger/interfaces/ILoggerFactory.ts
@@ -1,5 +1,5 @@
 import { injectable } from 'inversify';
-import { ILogger } from './logger';
+import { ILogger } from './ILogger';
 
 @injectable()
 export abstract class ILoggerFactory {
