+++ src/utils/packageInfo/index.ts
@@ -1,4 +1,4 @@
-import { IPackageInfo } from './interfaces/packageInfo';
+import { IPackageInfo } from './interfaces/IPackageInfo';
 import { PackageInfo } from './impl/packageInfo';
 import { interfaces } from 'inversify';
 import Bind = interfaces.Bind;
@@ -7,4 +7,4 @@ export const packageInfoModulesBinder = (bind: Bind): void => {
   bind<IPackageInfo>(IPackageInfo).to(PackageInfo).inSingletonScope();
 };
 
-export * from './interfaces/packageInfo';
+export * from './interfaces/IPackageInfo';
