+++ src/utils/packageInfo/interfaces/IPackageInfo.ts
