+++ src/utils/yarn/impl/yarn.ts
@@ -1,8 +1,8 @@
 import { ICommandRunner } from '../../commandRunner';
-import { IYarn, IYarnOptions } from '../interfaces/yarn';
+import { IYarn, IYarnOptions } from '../interfaces/IYarn';
 import { injectable } from 'inversify';
 import { ILoggerFactory } from '../../logger';
-import { ILogger } from '../../logger/interfaces/logger';
+import { ILogger } from '../../logger/interfaces/ILogger';
 
 @injectable()
 export class Yarn extends IYarn {
