+++ src/utils/yarn/index.ts
@@ -1,10 +1,10 @@
 import { interfaces } from 'inversify';
 import { Yarn } from './impl/yarn';
-import { IYarn } from './interfaces/yarn';
+import { IYarn } from './interfaces/IYarn';
 import Bind = interfaces.Bind;
 
 export const yarnModulesBinder = (bind: Bind): void => {
   bind<IYarn>(IYarn).to(Yarn).inSingletonScope();
 };
 
-export * from './interfaces/yarn';
+export * from './interfaces/IYarn';
