+++ test/src/utils/commandRunner/impl/commandRunner.spec.ts
@@ -2,7 +2,7 @@ import { CommandRunner } from '../../../../../src/utils/commandRunner/impl/comma
 // @ts-ignore
 import mockSpawn = require('mock-spawn');
 import { ILoggerFactory } from '../../../../../src/utils/logger';
-import { ILogger } from '../../../../../src/utils/logger/interfaces/logger';
+import { ILogger } from '../../../../../src/utils/logger/interfaces/ILogger';
 
 describe(`command runner`, () => {
   let mock = mockSpawn();
